Add a string class with related iterator and test
- Use the new string class in various parts of the engine
- Remove inclusion of cstring header where not needed
- Reorganize delete order in application shutdown method
- Convert iterators to constant ones in the list test
- Fix the include directories setting for MSVC in the CMake script
- Fix a set of warnings on Android regarding the EGL device class
- Fix an integer conversion warning in the PNG texture loader class

+++ src/android/EglGfxDevice.cpp
@@ -48,10 +48,10 @@ bool EglGfxDevice::isModeSupported(struct android_app *state, DisplayMode mode)
 	const EGLint attribs[] =
 	{
 		EGL_SURFACE_TYPE,		EGL_WINDOW_BIT,
-		EGL_BLUE_SIZE,			mode.blueBits(),
-		EGL_GREEN_SIZE,			mode.greenBits(),
-		EGL_RED_SIZE,			mode.redBits(),
-		EGL_ALPHA_SIZE,			mode.alphaBits(),
+		EGL_BLUE_SIZE,			static_cast<int>(mode.blueBits()),
+		EGL_GREEN_SIZE,			static_cast<int>(mode.greenBits()),
+		EGL_RED_SIZE,			static_cast<int>(mode.redBits()),
+		EGL_ALPHA_SIZE,			static_cast<int>(mode.alphaBits()),
 		EGL_NONE
 	};
 
@@ -91,10 +91,10 @@ void EglGfxDevice::initDevice(struct android_app* state)
 	const EGLint attribs[] =
 	{
 		EGL_SURFACE_TYPE,		EGL_WINDOW_BIT,
-		EGL_BLUE_SIZE,			mode_.blueBits(),
-		EGL_GREEN_SIZE,			mode_.greenBits(),
-		EGL_RED_SIZE,			mode_.redBits(),
-		EGL_ALPHA_SIZE,			mode_.alphaBits(),
+		EGL_BLUE_SIZE,			static_cast<int>(mode_.blueBits()),
+		EGL_GREEN_SIZE,			static_cast<int>(mode_.greenBits()),
+		EGL_RED_SIZE,			static_cast<int>(mode_.redBits()),
+		EGL_ALPHA_SIZE,			static_cast<int>(mode_.alphaBits()),
 		EGL_NONE
 	};
